Silence warnings on MacOS caused by off_t being a long long
One MacOS off_t is a long long int and not a long int so we need to
care about that in our format strings.

+++ src/keyring/keyring_file.c
@@ -73,7 +73,8 @@ get_key_by_name(GenericKeyring *keyring, const char *key_name, KeyringReturnCode
 					errcode_for_file_access(),
 					errmsg("keyring file \"%s\" is corrupted: %m",
 						   file_keyring->file_name),
-					errdetail("invalid key size %lu expected %lu", bytes_read, sizeof(KeyInfo)));
+					errdetail("invalid key size %lld expected %lu",
+							  (long long int) bytes_read, sizeof(KeyInfo)));
 			return NULL;
 		}
 		if (strncasecmp(key->name, key_name, sizeof(key->name)) == 0)
